04 auth

+++ src/controllers/notesController.js
@@ -4,7 +4,7 @@ import { Note } from '../models/note.js';
 export const getAllNotes = async (req, res) => {
   const { page = 1, perPage = 10, tag, search } = req.query;
   const skip = (page - 1) * perPage;
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({ userId: req.userId });
 
   if (tag) {
     notesQuery.where('tag').equals(tag);
@@ -32,7 +32,7 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
   const { noteId: _id } = req.params;
-  const note = await Note.findOne({ _id });
+  const note = await Note.findOne({ _id, userId: req.user._id });
   if (!note) {
     throw createHttpError(404, 'Note not found');
   }
@@ -40,13 +40,14 @@ export const getNoteById = async (req, res) => {
 };
 
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({ ...req.body, userId: req.user._id });
+
   res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res) => {
   const { noteId: _id } = req.params;
-  const note = await Note.findOneAndDelete({ _id });
+  const note = await Note.findOneAndDelete({ _id, userId: req.user._id });
   if (!note) {
     throw createHttpError(404, 'Note not found');
   }
@@ -55,9 +56,13 @@ export const deleteNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
   const { noteId: _id } = req.params;
-  const note = await Note.findOneAndUpdate({ _id }, req.body, {
-    returnDocument: 'after',
-  });
+  const note = await Note.findOneAndUpdate(
+    { _id, userId: req.user._id },
+    req.body,
+    {
+      returnDocument: 'after',
+    },
+  );
   if (!note) {
     throw createHttpError(404, 'Note not found');
   }
